backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
anthropic==0.40.0
chromadb==0.5.23
llama-index-core==0.11.23
llama-index-vector-stores-chroma==0.3.0
llama-index-embeddings-huggingface==0.3.0
pymupdf==1.24.11
python-docx==1.1.2
sentence-transformers==3.3.1
pydantic-settings==2.5.2
httpx==0.28.1
aiofiles==24.1.0
